bugs resolved

+++ database/migrations/2023_08_19_063422_add_type_column_to_team_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('teams', function (Blueprint $table) {
-            $table->double('type')->nullable()->after('language_id');
+            $table->integer('type')->nullable()->after('language_id');
         });
     }
